clean up notification_handler
